Improved memory usage
Removed useless bitarray to reduce call to GC ( huge pef improvement)
Added compiltion options for ARM


Former-commit-id: d68a524dbd95bc4c5c4a4e65c59c2b20145e1c09 [formerly e01f4fcf4117a34865a81b85b5a3e6d21090cc80]
Former-commit-id: e37a452055b0c79afb0175228119be48a1ddb130 [formerly 7347291c6911072de29b91c08dbb65947b258e40]
Former-commit-id: f396467ed048dcb852f696092ec02d4386283981
Former-commit-id: 047beae2413fa015c42d08884d39c39c32aee4de

+++ Source/RawParserUWP/RawParserUWP/Model/Format/Image/RawImage.cs
@@ -82,9 +82,10 @@ namespace RawParserUWP.Model.Format.Image
                     {
                         //get the pixel
                         ushort temp = 0;
+                        //todo check if correct
                         for (int k = 0; k < colorDepth; k++)
                         {
-                            if (imageData[(i * (int)colorDepth) + k])
+                            if (imageData[(i * colorDepth) + k])
                             {
                                 temp |= (ushort)(1 << k);
                             }
